fix: refactor cn utils
