+++ src/components/ui/form.tsx
@@ -13,7 +13,7 @@ import {
 } from 'react-hook-form';
 
 import { Label } from '@/components/ui/label';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib';
 
 const Form = FormProvider;
